backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
crewai
akshare>=1.18.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
langchain-openai
pydantic>=2.0.0
